+++ reflex-dom/ChangeLog.md
@@ -1,7 +1,7 @@
 # Revision history for reflex-dom
 
 ## Unreleased
-* On Android, enable propting the user for geolocation
+* On Android, enable prompting the user for geolocation
   permissions on demand by default. See
   https://developer.android.com/reference/android/webkit/WebChromeClient.html#onGeolocationPermissionsShowPrompt(java.lang.String,%20android.webkit.GeolocationPermissions.Callback)
   for details.