+++ reflex-dom/ChangeLog.md
@@ -1,10 +1,13 @@
 # Revision history for reflex-dom
 
 ## 0.5.3
+
 * On Android, enable prompting the user for geolocation
   permissions on demand by default. See
   https://developer.android.com/reference/android/webkit/WebChromeClient.html#onGeolocationPermissionsShowPrompt(java.lang.String,%20android.webkit.GeolocationPermissions.Callback)
   for details.
+
+ * Add `< 0.7` upper bound for reflex
 
 ## 0.5.2